Move stubs to NP
