+++ src/plugins/data/common/es_query/filters/stubs/index.ts
@@ -17,22 +17,7 @@
  * under the License.
  */
 
-import { esFilters } from '../../../../../../../../../plugins/data/public';
-
-export const phraseFilter: esFilters.PhraseFilter = {
-  meta: {
-    negate: false,
-    index: 'logstash-*',
-    type: 'phrase',
-    key: 'machine.os',
-    value: 'ios',
-    disabled: false,
-    alias: null,
-    params: {
-      query: 'ios',
-    },
-  },
-  $state: {
-    store: esFilters.FilterStateStore.APP_STATE,
-  },
-};
+export * from './exists_filter';
+export * from './phrase_filter';
+export * from './phrases_filter';
+export * from './range_filter';
